Comeback

+++ Assets/Scripts/Level/LevelRewardData.cs
@@ -1,7 +1,7 @@
 ﻿[System.Serializable]
 public class LevelRewardData
 {
-    public int seconds;
+    public float seconds;
     public int rewardAmount;
 }
